EA version for experiments
Planes-EA has been added, it contains a version of the EA that runs for a set no. of evals (budget=100000).

The class files that make up the model have been moved to PlaneModel.py to make the notebook easier to read.

Example_Case_Clear.xlsx is a copy of he problem instance with the cell formatting fixed so that it works with the notebook. If you have problems reading the sheet, switch to this copy.